[Shaan] | IoT Device Catalogue Refactor [2] (#21)
