+++ packages/frontend/src/App.tsx
@@ -13,9 +13,9 @@ import profileLoader from "./loaders/profileLoader";
 import Profile from "./pages/Profile";
 import Logout from "./pages/Logout";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import CreateInventory from "./pages/IoTDeviceCatalogue/CreateInventory";
-import ManageInventory from "./pages/IoTDeviceCatalogue/ManageInventory";
-import EditInventory from "./pages/IoTDeviceCatalogue/EditInventory";
+import CreateInventory from "./pages/CreateInventory";
+import ManageInventory from "./pages/ManageInventory";
+import EditInventory from "./pages/EditInventory";
 import DarkLightModeToggle from "./components/DarkLightModeToggle";
 
 const queryClient = new QueryClient();
