+++ packages/frontend/src/components/Navbar.tsx
@@ -19,7 +19,7 @@ import {
 import { Link } from "react-router-dom";
 import useMe from "../hooks/useMe";
 import { UserSchema } from "../api/generated";
-import logo from "../../public/icon.svg";
+import logo from "../assets/icon.svg";
 import { ChevronDownIcon } from "@chakra-ui/icons";
 
 // Path: packages\frontend\src\components\Navbar.tsx
