+++ packages/frontend/src/utils/currencyFormatter.ts
@@ -1,0 +1,8 @@
+export function convertToCurrency(price: number): string {
+  if (!price) return "N/A";
+
+  return price.toLocaleString("en-US", {
+    style: "currency",
+    currency: "USD",
+  });
+}
